Replace daily reports response
Daily reports previously sent as the response to GET requests are now replaced by a single monthly report. The new monthly report includes total values and the according percentage sold by each vendor, daily reports (these are the same as in the old response) and the total amount sold in the requested month.